+++ src/OrlovMikhail.GraphViz.Writing/Attributes/Height/HeightAttributeExtensions.cs
@@ -2,7 +2,7 @@
 {
     public static class HeightAttributeExtensions
     {
-        public static IAttrSet Height(this IAttrSet attrSet, double value)
+        public static IAttrSet Height(this IAttrSet attrSet, decimal value)
         {
             HeightAttribute a = new HeightAttribute(value);
             attrSet.Add(a);
