+++ src/OrlovMikhail.GraphViz.Writing/Attributes/NodeSep/RankSepAttribute.cs
@@ -2,7 +2,7 @@
 {
     public class RankSepAttribute : DoubleAttribute
     {
-        public RankSepAttribute(double value) : base(value)
+        public RankSepAttribute(decimal value) : base(value)
         {
         }
     }
