+++ src/OrlovMikhail.GraphViz.Writing/Attributes/RankSep/NodeSepAttribute.cs
@@ -2,7 +2,7 @@
 {
     public class NodeSepAttribute : DoubleAttribute
     {
-        public NodeSepAttribute(double value) : base(value)
+        public NodeSepAttribute(decimal value) : base(value)
         {
         }
     }
